Домашка по "4.2. Управление схемами баз данных"

+++ src/main/java/ru/hogwarts/school/repositories/AvatarRepository.java
@@ -6,8 +6,9 @@ import ru.hogwarts.school.model.Avatar;
 
 import java.util.Optional;
 @Repository
-public interface AvatarRepository extends PagingAndSortingRepository<Avatar, Integer> {
+public interface AvatarRepository extends PagingAndSortingRepository<Avatar, Long> {
 
     Optional<Avatar> findByStudentId(Long studentId);
+
 
 }
